improve splitter implementation and css classes
