+++ src/components/aggregations/aggregations.less
@@ -7,5 +7,4 @@
   background-color: @gray8;
   position: relative;
   min-height: 0;
-  flex-shrink: 1;
 }
