+++ src/components/pipeline-workspace/splitter.less
@@ -1,17 +1,9 @@
 .splitter {
   position: absolute;
-  top: 34px;
+  top: 0;
   left: 400px;
+  bottom: 0;
   border-right: 2px solid rgba(0, 0, 0, 0.1);
   box-shadow: 1px 0px 2px rgba(0, 0, 0, 0.1);
-  bottom: 0;
   z-index: 1;
-}
-
-.splitter-editing-collation {
-  top: 76px;
-}
-
-.splitter-editing-view {
-  top: 93px;
 }
