+++ src/components/pipeline/pipeline.less
@@ -2,11 +2,12 @@
 
 .pipeline {
   display: flex;
-  flex-shrink: 1;
+  flex-grow: 1;
   flex-direction: column;
   width: 100%;
   height: 100%;
   min-height: 0;
+  position: relative;
 
   &-fullscreen {
     position: absolute;
